server: fix bogus startup log message (#9404)
When Envoy splits the empty `--disable-extensions` flag at startup,
it ends up with a vector containing a single empty string. Since
the vector isn't empty, it tries to disable the extension with an
empty name. The fix is to not try to split the option string if it
is empty.

This fixes #9377.

Signed-off-by: James Peach <jpeach@apache.org>
Signed-off-by: Prakhar <prakhar_au@yahoo.com>